util/hlc: add protection against clock jumps
A cluster setting is added to panic on clock jumps. The existing
safeguard in forward clock jumps is in periodic inter-node heartbeats,
so there is a window where anomalies could occur.

This change adds forward and backward jump checks to HLC (and a background
goroutine to keep lastPhysicalTime up to date).

A cluster setting is added to periodically persist wall time to disk.
The HLC is guaranteed to not use a wall time greater than the persisted
value. When cockroach restarts, it waits for the wall time to reach the
persisted value before starting up. This guarantees monotonic wall time
across restarts

Fixes #23214

Release note (general change): Added cluster settings for HLC to be
monotonic across restarts and the server to panic on clock jumps